Changed behaviour of starred DataSource after editing it
* StarList allows access to isStarred for all 
With this change it is easier to check if any DataSource is starred/ a favourite and use this information.

* Suggestion to "Editing a starred DataSource"
With this change after editing a starred DataSource the changed DataSource will be starred and the old one loses the star. I think this is more intuitive than the old behaviour.